feat: inject rest periods into straight-set workout blocks
Split blocksFromMovementIds tests into dedicated file to stay under
the 400-line hard cap.
