+++ convex/coach/weekProgrammingHelpers.test.ts
@@ -1,6 +1,5 @@
 import { describe, expect, it } from "vitest";
 import {
-  blocksFromMovementIds,
   formatSessionTitle,
   inferArmPosition,
   type SessionType,
@@ -194,129 +193,5 @@ describe("sortForMinimalEquipmentSwitches", () => {
 
   it("handles empty input", () => {
     expect(sortForMinimalEquipmentSwitches([], catalogWithAccessories)).toEqual([]);
-  });
-});
-
-// ---------------------------------------------------------------------------
-// blocksFromMovementIds
-// ---------------------------------------------------------------------------
-
-const blockCatalog = [
-  {
-    id: "bench",
-    countReps: true,
-    onMachineInfo: { accessory: "Smart Bar" },
-  },
-  {
-    id: "row",
-    countReps: true,
-    onMachineInfo: { accessory: "Smart Bar" },
-  },
-  {
-    id: "curl",
-    countReps: true,
-    onMachineInfo: { accessory: "Handle" },
-  },
-  {
-    id: "fly",
-    countReps: true,
-    onMachineInfo: { accessory: "Handle" },
-  },
-  {
-    id: "extension",
-    countReps: true,
-    onMachineInfo: { accessory: "Handle" },
-  },
-  {
-    id: "pushdown",
-    countReps: true,
-    onMachineInfo: { accessory: "Rope" },
-  },
-  {
-    id: "pushup",
-    countReps: false,
-    // no onMachineInfo — bodyweight/duration-based
-  },
-];
-
-describe("blocksFromMovementIds", () => {
-  it("returns empty array for empty input", () => {
-    expect(blocksFromMovementIds([])).toEqual([]);
-  });
-
-  it("groups exercises by accessory into 2-exercise superset blocks", () => {
-    // Pre-sorted by accessory: Smart Bar x2, Handle x2
-    const ids = ["bench", "row", "curl", "fly"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    expect(blocks).toHaveLength(2);
-    expect(blocks[0].exercises.map((e) => e.movementId)).toEqual(["bench", "row"]);
-    expect(blocks[1].exercises.map((e) => e.movementId)).toEqual(["curl", "fly"]);
-  });
-
-  it("puts odd exercise in its own straight-set block", () => {
-    // Handle x3 = 2 in superset + 1 solo
-    const ids = ["curl", "fly", "extension"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    expect(blocks).toHaveLength(2);
-    expect(blocks[0].exercises).toHaveLength(2);
-    expect(blocks[1].exercises).toHaveLength(1);
-    expect(blocks[1].exercises[0].movementId).toBe("extension");
-  });
-
-  it("creates separate blocks per accessory group", () => {
-    // Smart Bar x1, Handle x1, Rope x1 = 3 solo blocks
-    const ids = ["bench", "curl", "pushdown"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    expect(blocks).toHaveLength(3);
-    expect(blocks[0].exercises[0].movementId).toBe("bench");
-    expect(blocks[1].exercises[0].movementId).toBe("curl");
-    expect(blocks[2].exercises[0].movementId).toBe("pushdown");
-  });
-
-  it("groups bodyweight exercises together", () => {
-    const ids = ["pushup", "curl"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    // pushup = bodyweight group (1 block), curl = Handle group (1 block)
-    expect(blocks).toHaveLength(2);
-    expect(blocks[0].exercises[0].movementId).toBe("pushup");
-    expect(blocks[1].exercises[0].movementId).toBe("curl");
-  });
-
-  it("uses duration for non-countReps exercises", () => {
-    const ids = ["pushup"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    expect(blocks[0].exercises[0].duration).toBe(30);
-    expect(blocks[0].exercises[0].reps).toBeUndefined();
-  });
-
-  it("applies suggested reps from progressive overload", () => {
-    const ids = ["bench"];
-    const suggestions = [{ movementId: "bench", suggestedReps: 8 }];
-    const blocks = blocksFromMovementIds(ids, suggestions, { catalog: blockCatalog });
-
-    expect(blocks[0].exercises[0].reps).toBe(8);
-  });
-
-  it("defaults to 3 sets and 10 reps without suggestions", () => {
-    const ids = ["bench"];
-    const blocks = blocksFromMovementIds(ids, undefined, { catalog: blockCatalog });
-
-    expect(blocks[0].exercises[0].sets).toBe(3);
-    expect(blocks[0].exercises[0].reps).toBe(10);
-  });
-
-  it("works without catalog (all exercises in one accessory group)", () => {
-    const ids = ["a", "b", "c", "d"];
-    const blocks = blocksFromMovementIds(ids);
-
-    // Without catalog, all are same accessory group (bodyweight fallback), paired into supersets
-    expect(blocks).toHaveLength(2);
-    expect(blocks[0].exercises.map((e) => e.movementId)).toEqual(["a", "b"]);
-    expect(blocks[1].exercises.map((e) => e.movementId)).toEqual(["c", "d"]);
   });
 });
